+++ examples/ios/swift-3.0/Backlink/AppDelegate.swift
@@ -43,7 +43,9 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         window?.makeKeyAndVisible()
 
         do {
-            try FileManager.default.removeItem(at: Realm.Configuration.defaultConfiguration.fileURL!)
+            if case let .file(url) = Realm.Configuration.defaultConfiguration.kind {
+                try FileManager.default.removeItem(at: url)
+            }
         } catch {}
 
         let realm = try! Realm()